multi_tool_agent/requirements.txt
llama_index
python-dotenv
requests
google-cloud-aiplatform>=1.38.0
google-adk==1.10.0
google-api-core>=2.11.0
google-generativeai>=0.3.0
vertexai